Update build script in package.json
Modified the "build" script in package.json to create an empty .nojekyll file in the output directory after the build process. This is necessary for correct rendering of the site on platforms like GitHub Pages, which use Jekyll by default and ignore certain files starting with underscore, causing problems with Next.js's routing. The .nojekyll file disables this behavior to ensure correct routing.