Remove Appveyor config and update GitHub workflow
The .appveyor.yml file was removed as we have migrated the CI/CD pipeline from Appveyor to GitHub Actions. The main.yml file under .github/workflows was subsequently extended to include 'BRANCH: gh-pages' in the publish_dir, aligning with the previously deleted Appveyor configuration. This ensures that our deployment practices remain consistent even with the transition to a new pipeline.